9.Implement_RAG_SYSTEM/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-core
langchain-chroma
faiss-cpu
pypdf